refactor: llama table filters
